refactor controller-manager cmd to align w/ upstream design
